Refactoring gcSelectors and gcAPI
