+++ src/components/Challenger.vue
@@ -25,7 +25,7 @@
 <script lang="ts">
 import { Options, Vue } from "vue-class-component";
 import { Ref, ref, watch } from "vue";
-import { gcSelectors } from "../utils/gcSelectors";
+import { gcSelectors } from "../utils/gc/selectors";
 import { cleanSelector } from "@/utils/StringUtils";
 import { FULL_LOBBY_PLAYERS_NUMBER } from "@/utils/magicNumbers";
 import serializer from "@/scripts/lobby/serializer";
