+++ src/scripts/lobby/LobbiesModifier.ts
@@ -6,7 +6,7 @@ import { createApp } from 'vue'
 import KDRComponent from '../../components/KDR.vue'
 import GCCMarkComponent from '../../components/GCCMarks.vue'
 import GCCLobbyTitle from '../../components/GCCLobbyTitle.vue'
-import { gcSelectors } from '../../utils/gcSelectors'
+import { gcSelectors } from '../../utils/gc/selectors'
 import serializer from './serializer'
 import lobbyFilter from './lobbyFilter'
 import Logger from 'js-logger'
