+++ src/scripts/lobby/MyLobbyModifier.ts
@@ -5,7 +5,7 @@ import { domEntityType } from './domain/domEntityType'
 import { createApp, reactive, ref } from 'vue'
 import KDRComponent from '../../components/KDR.vue'
 import GCChallengerComponent from '../../components/Challenger.vue'
-import { gcSelectors } from '../../utils/gcSelectors'
+import { gcSelectors } from '../../utils/gc/selectors'
 import serializer from './serializer'
 import lobbyFilter from './lobbyFilter'
 import Logger from 'js-logger'
