+++ src/scripts/preMatch/PreMatchModifier.ts
@@ -3,7 +3,7 @@ import $ from 'jquery'
 import { createApp } from 'vue'
 import KDRComponent from '../../components/KDR.vue'
 import GCCLogo from '../../components/GCCLogo.vue'
-import { gcSelectors } from '../../utils/gcSelectors'
+import { gcSelectors } from '../../utils/gc/selectors'
 import serializer from '../lobby/serializer'
 import Logger from 'js-logger'
 import BrowserStorage from '../../utils/storage'
