+++ src/scripts/team/TeamModifier.ts
@@ -4,7 +4,7 @@ import $ from 'jquery'
 import { domEntityType } from '../lobby/domain/domEntityType'
 import { createApp } from 'vue'
 import KDRComponent from '../../components/KDR.vue'
-import { gcSelectors } from '../../utils/gcSelectors'
+import { gcSelectors } from '../../utils/gc/selectors'
 import serializer from '../lobby/serializer'
 import LobbyPlayer from '../lobby/domain/LobbyPlayer'
 import BrowserStorage from '../../utils/storage'
