+++ src/utils/devAuxScripts/index.ts
@@ -1,7 +1,7 @@
 import '../../plugins/window.setup'
 import logger from 'js-logger'
 import serializer from '../../scripts/lobby/serializer'
-import { gcSelectors } from '../gcSelectors'
+import { gcSelectors } from '../gc/selectors'
 import { gcLevelsMap } from '../gc/levels'
 import { userAPI } from '../gc/api'
 import { gcAssetsUrls, gcUrls } from '../gcUrls'
